Add initial implementation to tail a file or stdin
As a first part to grab entries from audit.log, implement a `tail`
functionality. The approach chosen here is to follow the file and deliver
string lines via channel. The tail implementation is watching for the file
deletion and in case the file disappears or inode changes it breaks the read
loop and quits program. Intended behavior here is that program gracefully
restarts and starts from beginning when there's a new file to be read.