Start tail from EOF and fix dead-lock (#8)
In order to not repeate already recorded metrics, it makes sense to start from
the end of the file. While there, the read loop sleep was reduced to 10ms.

A read dead-lock bug was also fixed by moving the initial read inside the
spawned goroutine. When it was in parent function, it would block on channel
write before the function returned and therefore nothing could actually read
the channel so it got blocked.